chore: Moved the fixture-helpers file to the tests/helpers folder
