+++ tests/fixtures/ember-container-query-customizations/index.js
@@ -1,4 +1,4 @@
-import { convertToJson } from '../../fixture-helpers.js';
+import { convertToJson } from '../../helpers/fixture.js';
 
 const inputProject = convertToJson('ember-container-query-customizations/input');
 const outputProject = convertToJson('ember-container-query-customizations/output');
