+++ tests/fixtures/ember-container-query-javascript/index.js
@@ -1,4 +1,4 @@
-import { convertToJson } from '../../fixture-helpers.js';
+import { convertToJson } from '../../helpers/fixture.js';
 
 const inputProject = convertToJson('ember-container-query-javascript/input');
 const outputProject = convertToJson('ember-container-query-javascript/output');
