+++ tests/fixtures/ember-container-query-typescript/index.js
@@ -1,4 +1,4 @@
-import { convertToJson } from '../../fixture-helpers.js';
+import { convertToJson } from '../../helpers/fixture.js';
 
 const inputProject = convertToJson('ember-container-query-typescript/input');
 const outputProject = convertToJson('ember-container-query-typescript/output');
