+++ tests/fixtures/new-v1-addon-javascript/index.js
@@ -1,4 +1,4 @@
-import { convertToJson } from '../../fixture-helpers.js';
+import { convertToJson } from '../../helpers/fixture.js';
 
 const inputProject = convertToJson('new-v1-addon-javascript/input');
 const outputProject = convertToJson('new-v1-addon-javascript/output');
