+++ tests/fixtures/new-v1-addon-npm/index.js
@@ -1,4 +1,4 @@
-import { convertToJson } from '../../fixture-helpers.js';
+import { convertToJson } from '../../helpers/fixture.js';
 
 const inputProject = convertToJson('new-v1-addon-npm/input');
 const outputProject = convertToJson('new-v1-addon-npm/output');
