+++ tests/fixtures/new-v1-addon-pnpm/index.js
@@ -1,4 +1,4 @@
-import { convertToJson } from '../../fixture-helpers.js';
+import { convertToJson } from '../../helpers/fixture.js';
 
 const inputProject = convertToJson('new-v1-addon-pnpm/input');
 const outputProject = convertToJson('new-v1-addon-pnpm/output');
